api/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
pyjwt

